more explicit exports pt.1
