+++ packages/client/wallets/aa/src/blockchain/wallets/paymaster.ts
@@ -1,6 +1,4 @@
-import { usesGelatoBundler } from "@/utils/blockchain";
 import { PAYMASTER_RPC, PM_BASE_RPC, PM_BASE_SEPOLIA_RPC } from "@/utils/constants";
-import { logInputOutput } from "@/utils/log";
 import { createZeroDevPaymasterClient } from "@zerodev/sdk";
 import { Middleware } from "permissionless/actions/smartAccount";
 import { EntryPoint } from "permissionless/types/entrypoint";
@@ -8,13 +6,14 @@ import { http } from "viem";
 
 import { EVMBlockchainIncludingTestnet } from "@crossmint/common-sdk-base";
 
+import { usesGelatoBundler } from "../../utils/blockchain";
 import { getViemNetwork, getZeroDevProjectIdByBlockchain } from "../BlockchainNetworks";
 
 export function usePaymaster(chain: EVMBlockchainIncludingTestnet) {
     return !usesGelatoBundler(chain);
 }
 
-const getPaymasterRPC = logInputOutput((chain: EVMBlockchainIncludingTestnet) => {
+const getPaymasterRPC = (chain: EVMBlockchainIncludingTestnet) => {
     switch (chain) {
         case EVMBlockchainIncludingTestnet.BASE_SEPOLIA:
             return PM_BASE_SEPOLIA_RPC;
@@ -23,7 +22,7 @@ const getPaymasterRPC = logInputOutput((chain: EVMBlockchainIncludingTestnet) =>
         default:
             return PAYMASTER_RPC + getZeroDevProjectIdByBlockchain(chain) + "?paymasterProvider=STACKUP";
     }
-}, "getPaymasterRPC");
+};
 
 export function paymasterMiddleware({
     entryPoint,
