+++ packages/client/wallets/aa/src/utils/auth.ts
@@ -1,7 +1,7 @@
 import { logError } from "@/services/logging";
 
-import { errorToJSON } from "../types/Error";
 import { SCW_SERVICE } from "./constants";
+import { errorToJSON } from "./log";
 
 export const parseToken = (token: any) => {
     try {
