+++ packages/client/wallets/aa/src/utils/log.ts
@@ -144,3 +144,14 @@ function logInfoIfNotInLocalhost(message: string, context?: object) {
     }
     logInfo(message, context);
 }
+
+export function errorToJSON(error: Error | unknown) {
+    const errorToLog = error instanceof Error ? error : { message: "Unknown error", name: "Unknown error" };
+
+    if (!(errorToLog instanceof Error) && (errorToLog as any).constructor?.name !== "SyntheticBaseEvent") {
+        logError("ERROR_TO_JSON_FAILED", { error: errorToLog });
+        throw new Error("[errorToJSON] err is not instanceof Error nor SyntheticBaseEvent");
+    }
+
+    return JSON.parse(JSON.stringify(errorToLog, Object.getOwnPropertyNames(errorToLog)));
+}
